woroking on showing a topic, taglist and minor bug fixes

+++ backend/src/api/v1/util/jwt_impl.h
@@ -34,7 +34,7 @@ inline std::optional<Token> verifyJWT(const std::string &token,
 	}  
 }
 
-inline std::string signJWT(const int user_id, 
+inline std::string signJWT(const size_t user_id, 
                            const std::string &username,
                            const std::string &secret,
                            const std::string issuer = "arth") {
